created tests crate
