+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 resolver = "2"
 
-members = [ "atlas", "comms", "comms-macros","rusty-roads"]
+members = [ "atlas", "comms", "comms-macros","rusty-roads", "tests"]
 
 [workspace.lints.clippy]
 # all = { level = "deny", priority = 1 }
