+++ comms/src/lib.rs
@@ -2,7 +2,7 @@ pub mod comms_types;
 
 pub use bytes::Bytes;
 pub use comms_macros::Parquet;
-pub use comms_types::{FromParquet, ParquetParseError, ToParquet};
+pub use comms_types::{Parquet, ParquetParseError};
 
 pub mod exports {
     pub use arrow_array::RecordBatch;
