+++ tests/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "tests"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+comms = { version = "0.1.0", path = "../comms" }
+geo-types = "0.7.15"
+rand = "0.9.0"
+rusty-roads = { version = "0.1.0", path = "../rusty-roads" }
+
+[lints]
+workspace = true
